Prevent hovering on existing reactions from shifting buttons down
Prior to this commit, in some cases, when highlighting an existing reaction,
the add reaction button would shift down to the right.

This would happen in circumstances where not all 8 reactions
were present on a comment.  This was due to buttons with no reaction
count being hidden with `display: none`.

Then Primer's button highlight CSS would add a border on the right
of the highlighted element and remove the left border on the following element.
This ensured that the row of buttons would remain a constant width.

However when some buttons were hidden, a border was added to the highlighted
element without the removal of a border from another element (since it was hidden).

This resulted in the row of buttons gaining a pixel, shifting everything down.

Resolved by removing Primer's border expanding behavior.