Prevent non-colored emojis from being invisible when selected
Prior to this commit, if a user did not have a font
that contained colored emojis, the reactions will fallback to a simple grayscale emoji.

However when hovering over a emoji from the reactions popover,
it would become invisible as a result of some hover CSS styling the color to white.

This commit resolves this by explicitly setting the color.