Workaround for GitHub search API bug (fixes #119)
If an issue title contains a number that is surrounded by two spaces,
the `in:title` search qualifier is ignored, and the `number` property of
issues is also taken into account. Issues that are returned based on
said `number` property also always have a `score` of 100, which places
them before the actual issue.

To work around this limitation, the `og:title` issue lookup function now
iterates over all returned issues and returns the first one that exactly
matches the OpenGraph title.