Dynamic page size [non-breaking change]
Added size optional argument, equaling 25 if not specified so old users will not see any difference. This addresses the issue of extra-long comment sections, causing scrolling difficulties.

If the specified size is less than 16, the hidden-page size will increase exponentially until 16 is reached, avoiding problems with rate limiting.

If the specified size is larger or equal to 16, the hidden-page size will be equal to the specified size. E.g. if no size argument is passed in, the hidden-page size will default to 25, which is the same as before.