README: misleading lock-in statement
The documentation is wildly misleading. Everything in this project gets
locked behind a for-profit, closed source, proprietary service in
Microsoft GitHub. It’s misleading to say that this isn’t tracked and has
no lock-in, when it clearly does. We would be better off leveraging the
*decentralized* part of distributed version control systems (DVCS)
instead of centralizing around a service like Microsoft GitHub.

Feel free to close and fix the wording yourself. The rewording is
facetious on purpose, but there is truth to the sentiment. (Merge
requests always get more attention, and I’d email you a patch, but we’re
locked into this platform’s source control model)